Revert "Batch processing"
