+++ batch-processing/.gitignore
@@ -1,3 +1,1 @@
 outputs/
-*tfstate*
-*.terraform*
