+++ batch-processing/go.mod
@@ -3,14 +3,5 @@ module github.com/CodeYourFuture/immersive-go-course/batch-processing
 go 1.19
 
 require (
-	github.com/aws/aws-sdk-go v1.44.114
-	github.com/stretchr/testify v1.8.0
-	gopkg.in/gographics/imagick.v2 v2.6.2
-)
-
-require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/jmespath/go-jmespath v0.4.0 // indirect
-	github.com/pmezard/go-difflib v1.0.0 // indirect
-	gopkg.in/yaml.v3 v3.0.1 // indirect
+	gopkg.in/gographics/imagick.v2 v2.6.2 // indirect
 )
