--- batch-processing/inputs/large-input.csv
@@ -1,101 +1,0 @@
-url
-https://picsum.photos/id/0/5616/3744
-https://picsum.photos/id/1/5616/3744
-https://picsum.photos/id/10/2500/1667
-https://picsum.photos/id/100/2500/1656
-https://picsum.photos/id/1000/5626/3635
-https://picsum.photos/id/1001/5616/3744
-https://picsum.photos/id/1002/4312/2868
-https://picsum.photos/id/1003/1181/1772
-https://picsum.photos/id/1004/5616/3744
-https://picsum.photos/id/1005/5760/3840
-https://picsum.photos/id/1006/3000/2000
-https://picsum.photos/id/1008/5616/3744
-https://picsum.photos/id/1009/5000/7502
-https://picsum.photos/id/101/2621/1747
-https://picsum.photos/id/1010/5184/3456
-https://picsum.photos/id/1011/5472/3648
-https://picsum.photos/id/1012/3973/2639
-https://picsum.photos/id/1013/4256/2832
-https://picsum.photos/id/1014/6016/4000
-https://picsum.photos/id/1015/6000/4000
-https://picsum.photos/id/1016/3844/2563
-https://picsum.photos/id/1018/3914/2935
-https://picsum.photos/id/1019/5472/3648
-https://picsum.photos/id/102/4320/3240
-https://picsum.photos/id/1020/4288/2848
-https://picsum.photos/id/1021/2048/1206
-https://picsum.photos/id/1022/6000/3376
-https://picsum.photos/id/1023/3955/2094
-https://picsum.photos/id/1024/1920/1280
-https://picsum.photos/id/1025/4951/3301
-https://picsum.photos/id/1026/4621/3070
-https://picsum.photos/id/1027/2848/4272
-https://picsum.photos/id/1028/5184/3456
-https://picsum.photos/id/1029/4887/2759
-https://picsum.photos/id/103/2592/1936
-https://picsum.photos/id/1031/5446/3063
-https://picsum.photos/id/1032/2880/1800
-https://picsum.photos/id/1033/2048/1365
-https://picsum.photos/id/1035/5854/3903
-https://picsum.photos/id/1036/4608/3072
-https://picsum.photos/id/1037/5760/3840
-https://picsum.photos/id/1038/3914/5863
-https://picsum.photos/id/1039/6945/4635
-https://picsum.photos/id/104/3840/2160
-https://picsum.photos/id/1040/4496/3000
-https://picsum.photos/id/1041/5184/2916
-https://picsum.photos/id/1042/3456/5184
-https://picsum.photos/id/1043/5184/3456
-https://picsum.photos/id/1044/4032/2268
-https://picsum.photos/id/1045/3936/2624
-https://picsum.photos/id/1047/3264/2448
-https://picsum.photos/id/1048/5616/3744
-https://picsum.photos/id/1049/3900/3120
-https://picsum.photos/id/1050/6000/4000
-https://picsum.photos/id/1051/4928/3264
-https://picsum.photos/id/1052/4000/2667
-https://picsum.photos/id/1053/3596/2393
-https://picsum.photos/id/1054/3079/1733
-https://picsum.photos/id/1055/5472/3648
-https://picsum.photos/id/1056/3988/2720
-https://picsum.photos/id/1057/6016/4016
-https://picsum.photos/id/1058/4608/3072
-https://picsum.photos/id/1059/7360/4912
-https://picsum.photos/id/106/2592/1728
-https://picsum.photos/id/1060/5598/3732
-https://picsum.photos/id/1061/3264/2448
-https://picsum.photos/id/1062/5092/3395
-https://picsum.photos/id/1063/4867/3215
-https://picsum.photos/id/1064/4236/2819
-https://picsum.photos/id/1065/3744/5616
-https://picsum.photos/id/1066/2144/1424
-https://picsum.photos/id/1067/5760/3840
-https://picsum.photos/id/1068/7117/4090
-https://picsum.photos/id/1069/3500/2333
-https://picsum.photos/id/107/5760/3840
-https://picsum.photos/id/1070/5472/3648
-https://picsum.photos/id/1071/3000/1996
-https://picsum.photos/id/1072/3872/2592
-https://picsum.photos/id/1073/5472/3648
-https://picsum.photos/id/1074/5472/3648
-https://picsum.photos/id/1075/4858/3239
-https://picsum.photos/id/1076/4835/3223
-https://picsum.photos/id/1077/3000/1995
-https://picsum.photos/id/1078/3000/2000
-https://picsum.photos/id/1079/4496/3000
-https://picsum.photos/id/108/2000/1333
-https://picsum.photos/id/1080/6858/4574
-https://picsum.photos/id/1081/5512/3708
-https://picsum.photos/id/1082/5416/3611
-https://picsum.photos/id/1083/5472/3648
-https://picsum.photos/id/1084/4579/3271
-https://picsum.photos/id/109/4287/2392
-https://picsum.photos/id/11/2500/1667
-https://picsum.photos/id/110/5616/3744
-https://picsum.photos/id/111/4400/2656
-https://picsum.photos/id/112/4200/2800
-https://picsum.photos/id/113/4168/2464
-https://picsum.photos/id/114/3264/2448
-https://picsum.photos/id/115/1500/1000
-https://picsum.photos/id/116/3504/2336
