+++ batch-processing/main_test.go
@@ -2,9 +2,9 @@ package main
 
 import (
 	"errors"
+	"reflect"
 	"testing"
 
-	"github.com/stretchr/testify/require"
 	"gopkg.in/gographics/imagick.v2/imagick"
 )
 
@@ -16,7 +16,9 @@ func TestGrayscaleMockError(t *testing.T) {
 	}
 
 	err := c.Grayscale("input.jpg", "output.jpg")
-	require.Error(t, err)
+	if err == nil {
+		t.Fatal("expected error")
+	}
 }
 
 func TestGrayscaleMockCall(t *testing.T) {
@@ -33,6 +35,10 @@ func TestGrayscaleMockCall(t *testing.T) {
 	}
 
 	err := c.Grayscale("input.jpg", "output.jpg")
-	require.NoError(t, err)
-	require.Exactly(t, expected, args)
+	if err != nil {
+		t.Fatal(err)
+	}
+	if !reflect.DeepEqual(expected, args) {
+		t.Fatalf("incorrect arguments: expected %v, got %v", expected, args)
+	}
 }
