--- batch-processing/terraform/data.tf
@@ -1,1 +1,0 @@
-data "aws_caller_identity" "current" {}
