--- batch-processing/terraform/iam.tf
@@ -1,56 +1,0 @@
-resource "aws_iam_policy" "S3ReadWriteGoCourse" {
-  name = "S3ReadWriteGoCourse"
-  path = "/"
-
-  policy = <<EOF
-{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Sid": "ListObjectsInBucket",
-      "Effect": "Allow",
-      "Action": ["s3:ListBucket"],
-      "Resource": ["arn:aws:s3:::${aws_s3_bucket.b.id}"]
-    },
-    {
-      "Sid": "AllObjectActions",
-      "Effect": "Allow",
-      "Action": "s3:*Object",
-      "Resource": ["arn:aws:s3:::${aws_s3_bucket.b.id}/*"]
-    }
-  ]
-}
-EOF
-}
-
-
-resource "aws_iam_role" "GoCourseLambdaUserReadWriteS3" {
-  name = "GoCourseLambdaUserReadWriteS3"
-
-  assume_role_policy = <<EOF
-{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Effect": "Allow",
-      "Principal": {
-        "AWS": "arn:aws:iam::${data.aws_caller_identity.current.id}:root"
-      },
-      "Action": "sts:AssumeRole"
-    },
-    {
-      "Effect": "Allow",
-      "Principal": {
-        "Service": "lambda.amazonaws.com"
-      },
-      "Action": "sts:AssumeRole"
-    }
-  ]
-}
-EOF
-}
-
-resource "aws_iam_role_policy_attachment" "GoCourseLambdaUserReadWriteS3" {
-  role       = aws_iam_role.GoCourseLambdaUserReadWriteS3.name
-  policy_arn = aws_iam_policy.S3ReadWriteGoCourse.arn
-}
